requirement.txt
asgiref==3.8.1
biopython==1.85
Django==5.2.1
numpy==2.2.6
sqlparse==0.5.3
